Correcting Reporting during failed runs
Previously elk_report used metrics array to determine the correct
global run status to display (read comments for further details).
It still does, however during a failed run the metrics[events]
array is not populated.  Added check to use self.status (from the
report transaction) in the event that their was a failure.